chore(build): migrate ESM build from Rollup to Rolldown
Replace the Rollup-based ESM build (build:esm) with Rolldown 1.x.
node resolution and TypeScript transform are handled natively by
Rolldown (Oxc), so @rollup/plugin-node-resolve and
@rollup/plugin-typescript are dropped; @rollup/plugin-replace maps to
the built-in replacePlugin and rollup-plugin-delete to an rmSync call.

The banner is injected in generateBundle to stay byte-identical, since
Rolldown normalizes comment blocks (CRLF->LF, strips leading space)
during rendering.

Output verified against the previous Rollup build: 177/177 files with
identical banners and semantically identical (minified) code.

Closes #4165
