+++ package.json
@@ -38,7 +38,7 @@
 		"start:plugin": "cross-env PLUGIN=true webpack-dev-server --open",
 		"build": "pnpm run build:production && pnpm run build:packaged && pnpm run build:theme && pnpm run build:plugin && pnpm run build:esm && pnpm run build:cjs && pnpm run build:plugin:types",
 		"build:cjs": "node ./config/cjs.js",
-		"build:esm": "rollup -c ./config/rollup/esm.js",
+		"build:esm": "rolldown -c ./config/rolldown/esm.js",
 		"build:production": "cross-env NODE_ENV=production webpack",
 		"build:packaged": "cross-env NODE_ENV=packaged webpack",
 		"build:dev": "cross-env NODE_ENV=development webpack",
@@ -117,9 +117,6 @@
 		"@commitlint/config-conventional": "^21.1.0",
 		"@eslint/js": "^10.0.1",
 		"@playwright/test": "^1.61.1",
-		"@rollup/plugin-node-resolve": "^16.0.3",
-		"@rollup/plugin-replace": "^6.0.3",
-		"@rollup/plugin-typescript": "^12.3.0",
 		"@semantic-release/changelog": "^6.0.3",
 		"@semantic-release/commit-analyzer": "^13.0.1",
 		"@semantic-release/exec": "^7.1.0",
@@ -155,8 +152,7 @@
 		"mini-css-extract-plugin": "^2.10.2",
 		"playwright": "^1.61.1",
 		"regenerator-runtime": "^0.14.1",
-		"rollup": "^4.62.2",
-		"rollup-plugin-delete": "^3.0.2",
+		"rolldown": "^1.1.5",
 		"sass": "^1.101.0",
 		"sass-loader": "^17.0.0",
 		"semantic-release": "^25.0.5",
